Added parameter validation to EventRegister

+++ src/main/java/no/ntnu/idatt1002/demo/repo/EventRegister.java
@@ -2,10 +2,9 @@ package no.ntnu.idatt1002.demo.repo;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 import no.ntnu.idatt1002.demo.dao.DAO;
 import no.ntnu.idatt1002.demo.data.Event;
+import no.ntnu.idatt1002.demo.util.VerifyInput;
 
 
 /**
@@ -20,20 +19,12 @@ public class EventRegister {
    * Constructor for the EventRegister class.
    *
    * @param dao the data access object
+   * @throws IllegalArgumentException if dao is null
    */
   public EventRegister(DAO dao) {
+    VerifyInput.verifyNotNull(dao, "dao");
     this.dao = dao;
     this.events = new ArrayList<>();
-  }
-
-  /**
-   * This method returns the events in the register in the form of lists.
-   *
-   * @return the events in the register as lists of strings
-   */
-  public Map<Integer, List<String>> getEvents() {
-    return events.stream()
-        .collect(Collectors.toMap(Event::getId, Event::getAttributes));
   }
 
   /**
@@ -41,7 +32,7 @@ public class EventRegister {
    */
   public void getAllEvents() {
     events = new ArrayList<>();
-    List<List<String>> events = dao.getAllFromTable("Recipe", "Event", "recipe_id");
+    List<List<String>> events = dao.getAllFromTable("Event", "Recipe", "recipe_id");
     packageToEvent(events);
   }
 
@@ -62,7 +53,7 @@ public class EventRegister {
    * @param id the id of the event
    * @return the index of the event
    */
-  private int getIdFromEvent(int id) {
+  private int getIndexFromEvent(int id) {
     for (int i = 0; i < events.size(); i++) {
       if (events.get(i).getId() == id) {
         return i;
@@ -75,9 +66,13 @@ public class EventRegister {
    * This method deletes an event from the database.
    *
    * @param id the id of the event to be deleted
+   * @throws IllegalArgumentException if the event does not exist
    */
   public void deleteEvent(int id) {
-    int index = getIdFromEvent(id);
+    int index = getIndexFromEvent(id);
+    if (index == -1) {
+      throw new IllegalArgumentException("Event with id " + id + " does not exist.");
+    }
     dao.deleteFromDatabase(events.get(index));
   }
 
@@ -87,8 +82,12 @@ public class EventRegister {
    * @param event_id the id of the event
    * @param recipe_id the id of the recipe
    * @param date the date of the event
+   * @throws IllegalArgumentException if the event does not exist
    */
   public void updateEvent(int event_id, int recipe_id, int date) {
+    if (getIndexFromEvent(event_id) == -1) {
+      throw new IllegalArgumentException("Event with id " + event_id + " does not exist.");
+    }
     Event event = new Event(event_id, recipe_id, date);
     dao.updateDatabase(event);
   }
@@ -99,6 +98,7 @@ public class EventRegister {
    * @param date the date of the event
    */
   public void getEventsByDate(int date) {
+    VerifyInput.verifyDateLength(date, "date");
     events = new ArrayList<>();
     List<List<String>> events = dao.filterFromTable(
         "Event", "date", Integer.toString(date), "recipe", "recipe_id");
@@ -114,15 +114,24 @@ public class EventRegister {
     List<Event> newEvents = new ArrayList<>();
     for (List<String> event : eventList) {
       newEvents.add(new Event(
-          Integer.parseInt(event.get(4)),
-          Integer.parseInt(event.get(5)),
-          event.get(1),
-          Integer.parseInt(event.get(6)),
-          event.get(3),
-          Integer.parseInt(event.get(2))
+          Integer.parseInt(event.get(0)),
+          Integer.parseInt(event.get(1)),
+          event.get(4),
+          Integer.parseInt(event.get(2)),
+          event.get(6),
+          Integer.parseInt(event.get(5))
       ));
     }
     this.events = newEvents;
+  }
+
+  /**
+   * This method returns the events.
+   *
+   * @return the events
+   */
+  public List<Event> getEvents() {
+    return events;
   }
 
   /**
